Move fromDir para um arquivo isolado
